Fix bug in event broadcasting (#12)
When a component is added or removed from an entity, the system list
currently being iterated over may no longer match the target entity. To
fix this, events are now only handled once the current event has
finished broadcasting. Still not entirely sure that this is the most
desirable behavior though.

The `EventPtr` and `EventMut` interface also had an update.