Split events into `GlobalEvent` and `TargetedEvent`, redesign targeted events. (#55)
- Split events into distinct `GlobalEvent` and `TargetedEvent` types.
This was pretty much already the case internally, but attempting to hide
this separation created some suboptimal code. `const` panics are no
longer necessary in the API, for instance.
- Simplify targeted events. Targeted events no longer carry a duplicate
`EntityId`, which reduces the event size a bit. Users must now write
`sender.send_to(target, Event)` instead of `sender.send(Event(target))`.
Performance should be improved, but there was not an existing benchmark
set up to measure this.
- Added targeted event benchmark.
- Combined tutorial into a single page.
- Updated all documentation.