Fix `#[derive(Event)]` on generic types with type params.
The problem is the `This<'a>` associated type requires the event to
outlive `'a`, but there's no way to actually specify that in the impl.
As a fix, we add `'static` bounds to all generic params.

AFAICT this isn't truly fixable unless Rust adds higher-kinded
polymorphism, which isn't happening any time soon. Alternatively, Rust
could remove the `'static` requirement from `TypeId::of` which would let
us delete `This`.